refactor: TagSecton, 공용 컴포넌트화 및 페이지 적용(그리고 약간의 디자인수정)
