+++ src/components/shared/TagSection.tsx
@@ -1,50 +1,62 @@
+import { Tags } from "@/components/ui/Tags";
 import { TAG_GROUPS } from "@/lib/constants/constants";
 
 type TagSectionProps = {
+  title: string; 
   tags: string[];
   selectedCategory: string | null;
   onChangeCategory: (category: string) => void;
   toggleTagSelector: (tag: string, allTags: string[], max: number) => void;
 };
 
-const TagSection = ({ tags, selectedCategory, onChangeCategory, toggleTagSelector }: TagSectionProps) => {
+const TagSection = ({ title, tags, selectedCategory, onChangeCategory, toggleTagSelector }: TagSectionProps) => {
   const selectedGroup = TAG_GROUPS.find((group) => group.key === selectedCategory);
 
   return (
     <div className="space-y-2 pt-10">
       <div className="flex items-center justify-between">
         <div className="flex items-center gap-1">
-          <span className="text-title2 font-bold text-text-04">게시물 주제를 선택해주세요.</span>
-          <span className="text-title2 font-bold text-primary-default">*</span>
+          <p className="text-title2 font-bold text-text-04">{title}</p>
+          <p className="text-title2 font-bold text-primary-default">*</p>
         </div>
-        <span className="text-sm text-body text-text-04">{tags.length} / 7</span>
+        <p className="text-sm text-body text-text-04">{tags.length} / 7</p>
       </div>
-
+  
       {/* 주제 선택 버튼 */}
       <div className="flex flex-wrap gap-2">
         {TAG_GROUPS.map((group) => (
           <button
             key={group.key}
             onClick={() => onChangeCategory(group.key)}
-            className={`rounded-lg px-4 py-2 ${selectedCategory === group.key ? "bg-gray-300" : "bg-gray-100"}`}
+            className="focus:outline-none"
           >
-            {group.title}
+            <Tags
+              label={group.title}
+              variant={selectedCategory === group.key ? "gray" : "grayLine"}
+              size="md"
+              className={`${selectedCategory === group.key ? "border-line-02" : ""} cursor-pointer`}
+            />
           </button>
         ))}
       </div>
-
+  
       {/* 선택된 카테고리의 태그 */}
       {selectedGroup && (
         <div className="space-y-2">
-          <p className="pt-10 font-bold">태그</p>
+          <p className="text-title2 pt-10 font-bold">태그</p>
           <div className="flex flex-wrap gap-2">
             {selectedGroup.tags.map((tag) => (
               <button
                 key={tag}
                 onClick={() => toggleTagSelector(tag, selectedGroup.tags, selectedGroup.max)}
-                className={`rounded-lg px-3 py-1 ${tags.includes(tag) ? "bg-gray-300" : "bg-gray-100"}`}
+                className="focus:outline-none"
               >
-                {tag}
+                <Tags
+                  label={tag}
+                  variant={tags.includes(tag) ? "gray" : "grayLine"}
+                  size="md"
+                  className={`${tags.includes(tag) ? "border-line-02" : ""} cursor-pointer`}
+                />
               </button>
             ))}
           </div>
